.history/backend/requirements_20250302150845.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
Pillow==10.0.0
xhtml2pdf==0.2.11
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
